Changed the Google image search requests to use cors-anywhere, a dedicated CORS server, instead of our own API route in Flask.
This is because using our own Flask server is too slow
because Flask can't do asynchronous requests,
so it has to wait for the whole Google request
before starting the next request.
By using a dedicated CORS server like cors-anywhere,
we can still make CORS requests to google.com,
which doesn't allow CORS requests,
and at the same time, still do async requests,
meaning the latency is still the same,
but the throughput is much, much higher.

I still left the Python code for this in case,
but we should use the cors-anywhere version,
since it's so much faster.